created booking api model and refactor api folders
