+++ compass_app/app/lib/config/dependencies.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import '../data/repositories/auth/auth_repository.dart';
 import '../data/repositories/auth/auth_repository_dev.dart';
 import '../data/repositories/auth/auth_repository_remote.dart';
-import '../data/services/auth_api_client.dart';
+import '../data/services/api/auth_api_client.dart';
 import '../data/services/shared_preferences_service.dart';
 import '../data/repositories/activity/activity_repository.dart';
 import '../data/repositories/activity/activity_repository_local.dart';
@@ -17,7 +17,7 @@ import '../data/repositories/destination/destination_repository_local.dart';
 import '../data/repositories/destination/destination_repository_remote.dart';
 import '../data/repositories/itinerary_config/itinerary_config_repository.dart';
 import '../data/repositories/itinerary_config/itinerary_config_repository_memory.dart';
-import '../data/services/api_client.dart';
+import '../data/services/api/api_client.dart';
 import '../domain/components/booking/booking_create_component.dart';
 import '../domain/components/booking/booking_share_component.dart';
