+++ compass_app/app/pubspec.yaml
@@ -15,9 +15,11 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_svg: ^2.0.10+1
+  freezed_annotation: ^2.4.4
   go_router: ^14.2.0
   google_fonts: ^6.2.1
   intl: any
+  json_annotation: ^4.9.0
   logging: ^1.2.0
   provider: ^6.1.2
   share_plus: ^7.2.2
@@ -31,6 +33,9 @@ dev_dependencies:
   mocktail: ^1.0.4
   integration_test:
     sdk: flutter
+  build_runner: ^2.4.12
+  freezed: ^2.5.7
+  json_serializable: ^6.8.0
 
 flutter:
   uses-material-design: true
