+++ compass_app/app/testing/fakes/services/fake_auth_api_client.dart
@@ -1,4 +1,4 @@
-import 'package:compass_app/data/services/auth_api_client.dart';
+import 'package:compass_app/data/services/api/auth_api_client.dart';
 import 'package:compass_app/utils/result.dart';
 import 'package:compass_model/src/model/auth/login_request/login_request.dart';
 import 'package:compass_model/src/model/auth/login_response/login_response.dart';
